+++ rusk/src/lib/http/chain/graphql/archive/finalized_block.rs
@@ -6,8 +6,8 @@
 
 //! Module for GraphQL that only pertains to finalized blocks.
 
-use crate::http::chain::graphql::{DBContext, OptResult};
-use async_graphql::{Context, FieldError, FieldResult, Object};
+use crate::http::chain::graphql::DBContext;
+use async_graphql::{Context, FieldError, FieldResult};
 
 /// Check if a block height matches a block hash for a finalized block.
 pub async fn check_finalized_block(
@@ -21,15 +21,4 @@ pub async fn check_finalized_block(
         .match_finalized_block_height_hash(block_height, &hex_block_hash)
         .await
         .map_err(|e| FieldError::new(format!("Cannot check block: {}", e)))
-}
-
-/// Get the last finalized block.
-pub async fn last_finalized_block(
-    ctx: &Context<'_>,
-) -> FieldResult<(u64, String)> {
-    let (_, archive) = ctx.data::<DBContext>()?;
-
-    archive.fetch_last_finalized_block().await.map_err(|e| {
-        FieldError::new(format!("Cannot get last finalized block: {}", e))
-    })
 }