+++ rusk/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add simulate transaction API [#1225]
+
+### Changed
+
+- Change plonk verification to use embed verification data by default [#3507]
+- Change responses for moonlight gql endpoints (archive node) [#3512]
+
 ## [1.1.1] - 2025-02-21
 
 ### Changed
@@ -345,6 +354,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add build system that generates keys for circuits and caches them.
 
 <!-- Issues -->
+[#3512]: https://github.com/dusk-network/rusk/issues/3512
+[#3507]: https://github.com/dusk-network/rusk/issues/3507
 [#3494]: https://github.com/dusk-network/rusk/issues/3494
 [#3481]: https://github.com/dusk-network/rusk/issues/3481
 [#3359]: https://github.com/dusk-network/rusk/issues/3359
@@ -364,6 +375,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1603]: https://github.com/dusk-network/rusk/issues/1603
 [#1371]: https://github.com/dusk-network/rusk/issues/1371
 [#1257]: https://github.com/dusk-network/rusk/pull/1257
+[#1225]: https://github.com/dusk-network/rusk/issues/1225
 [#1219]: https://github.com/dusk-network/rusk/issues/1219
 [#1144]: https://github.com/dusk-network/rusk/issues/1144
 [#1080]: https://github.com/dusk-network/rusk/issues/1080